txn & producer: remove two TODOs
For txn, the TODO is dependent on the work being supported in Kafka: the
TODO has existed for a long time now and there's nothing that can be
done because Kafka is not supporting it yet.

For producer, not handling concurrent transactions in init producer ID
has never caused problems. It seems rare / likely to happen on client
initialization if it ever does.