consuming: allow purging consume topics
This is the first part in a series that allows users to cleanup
resources on long lived clients, as well as to add more consuming to
long lived clients.

Here, we allow the user to purge resources from the client for topics
the user is no longer interested in ever consuming.

This finally addresses two long standing TODOs in the client which were
only ever relevant if we introduced the concept to lose assignments.

As well, this also fixes a minor thing where the client would track
regular expressions as topic names (although these were not requested).